job/requirements.txt
psycopg2-binary
pymongo
pandas
azure-storage-blob
openpyxl
dotenv